updated to work with PowerShell 3.0; fixed some errors that prevented parsing from working correctly

+++ Tests/Rules/DscTestsPresent.tests.ps1
@@ -3,36 +3,38 @@ Import-Module -Verbose PSScriptAnalyzer
 $currentPath = Split-Path -Parent $MyInvocation.MyCommand.Path
 $ruleName = "PSDSCDscTestsPresent"
 
-Describe "DscTestsPresent rule in class based resource" {
+if ($PSVersionTable.PSVersion -ge [Version]'5.0') {
+    Describe "DscTestsPresent rule in class based resource" {
     
-    $testsPath = "$currentPath\DSCResources\MyDscResource\Tests"
-    $classResourcePath = "$currentPath\DSCResources\MyDscResource\MyDscResource.psm1"
+        $testsPath = "$currentPath\DSCResources\MyDscResource\Tests"
+        $classResourcePath = "$currentPath\DSCResources\MyDscResource\MyDscResource.psm1"
 
-    Context "When tests absent" {
+        Context "When tests absent" {
         
-        $violations = Invoke-ScriptAnalyzer -ErrorAction SilentlyContinue $classResourcePath | Where-Object {$_.RuleName -eq $ruleName}
-        $violationMessage = "No tests found for resource 'FileResource'"
+            $violations = Invoke-ScriptAnalyzer -ErrorAction SilentlyContinue $classResourcePath | Where-Object {$_.RuleName -eq $ruleName}
+            $violationMessage = "No tests found for resource 'FileResource'"
 
-        It "has 1 missing test violation" {
-            $violations.Count | Should Be 1
+            It "has 1 missing test violation" {
+                $violations.Count | Should Be 1
+            }
+
+            It "has the correct description message" {
+                $violations[0].Message | Should Be $violationMessage
+            }
         }
 
-        It "has the correct description message" {
-            $violations[0].Message | Should Be $violationMessage
+        Context "When tests present" {  
+            New-Item -Path $testsPath -ItemType Directory
+            New-Item -Path "$testsPath\FileResource_Test.psm1" -ItemType File
+
+            $noViolations = Invoke-ScriptAnalyzer -ErrorAction SilentlyContinue $classResourcePath | Where-Object {$_.RuleName -eq $ruleName}
+
+            It "returns no violations" {
+                $noViolations.Count | Should Be 0
+            }
+
+            Remove-Item -Path $testsPath -Recurse -Force
         }
-    }
-
-    Context "When tests present" {  
-        New-Item -Path $testsPath -ItemType Directory
-        New-Item -Path "$testsPath\FileResource_Test.psm1" -ItemType File
-
-        $noViolations = Invoke-ScriptAnalyzer -ErrorAction SilentlyContinue $classResourcePath | Where-Object {$_.RuleName -eq $ruleName}
-
-        It "returns no violations" {
-            $noViolations.Count | Should Be 0
-        }
-
-        Remove-Item -Path $testsPath -Recurse -Force
     }
 }
